requirements.txt
aiogram==3.17.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0